Refactor imports of types related to demographic models.
* Fix unresolved links in docs
