--- doc/pages/demographic_models.md
@@ -1,56 +1,0 @@
-(demographic-models)=
-
-# The demography debugger
-
-```{eval-rst}
-.. autoclass:: fwdpy11.DemographyDebugger
-   :members:
-```
-
-# Pre-computed demographic models
-
-```{eval-rst}
-.. automodule:: fwdpy11.demographic_models
-   :members:
-```
-
-```{eval-rst}
-.. autoclass:: fwdpy11.demographic_models.DemographicModelDetails
-   :members:
-```
-
-```{eval-rst}
-.. autoclass:: fwdpy11.demographic_models.DemographicModelCitation
-```
-
-## `fwdpy11.demographic_models.IM`
-
-```{eval-rst}
-.. automodule:: fwdpy11.demographic_models.IM
-```
-
-```{eval-rst}
-.. autofunction:: fwdpy11.demographic_models.IM.two_deme_IM
-```
-
-```{eval-rst}
-.. autoclass:: fwdpy11.demographic_models.IM.TwoDemeIMParameters
-```
-
-```{eval-rst}
-.. autoclass:: fwdpy11.demographic_models.IM.TwoDemeIMMetaData
-```
-
-## Human demographic models
-
-This module provides the following models:
-
-* {cite}`Tennessen2012-gg` via {func}`fwdpy11.demographic_models.human.tennessen`.
-* The three deme model from Table 2 of {cite}`Jouganous2017-tg` via {func}`fwdpy11.demographic_models.human.jouganous_three_deme`.
-
-```{eval-rst}
-.. automodule:: fwdpy11.demographic_models.human
-   :members:
-```
-
-
