+++ doc/short_vignettes/demes_vignette.md
@@ -51,7 +51,7 @@ Be sure to read the documentation for {func}`fwdpy11.discrete_demography.from_de
 There are important options concerning the run time of the simulation, etc.
 :::
 
-The return value is an instance of {class}`fwdpy11.demographic_models.DemographicModelDetails` and may be passed as the `demography` keyword argument to initialize an instance of {class}`fwdpy11.ModelParams`.
+The return value is an instance of {class}`fwdpy11.DemographicModelDetails` and may be passed as the `demography` keyword argument to initialize an instance of {class}`fwdpy11.ModelParams`.
 To extract the simulation length to generate the `simlen` parameter of your {class}`fwdpy11.ModelParams` instance:
 
 ```{code-cell} python
