+++ fwdpy11/demographic_models/__init__.py
@@ -1,3 +1,1 @@
 from . import human
-from .demographic_model_details import (DemographicModelCitation,  # NOQA
-                                        DemographicModelDetails)
